Continuação UC4.0 (Vendas)

+++ Codigo Fonte/DSC.JROficina.Aplicacao/src/DSC/JROficina/Aplicacao/Compra.java
@@ -87,7 +87,7 @@ public class Compra implements Entidade{
         return valor;
     }
 
-    public void setValor(float valor) {
+    public void setValor(double valor) {
         this.valor = valor;
     }
